refactor: Test 코드 내 도메인 로직 Class 분리 리팩토링
